polychem-ai_backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
rdkit
langchain
langchain-google-genai
python-dotenv
diskcache==5.6.3
